Added useAuth custom hook, also implemented in navbar for login/logout purposes
