+++ src/components/Navbar/Navbar.jsx
@@ -4,12 +4,12 @@ import { useState, useEffect } from "react";
 import "./Navbar.css";
 import Link from "next/link";
 import Image from "next/image";
-import supabase from "@/utils/supabase/client";
+import useAuth from "@/hooks/useAuth";
 
 export default function Navbar() {
+  const { user, logout } = useAuth();
   const [isVisible, setIsVisible] = useState(true);
   const [lastScrollY, setLastScrollY] = useState(0);
-  const [user, setUser] = useState(null);
 
   const handleScroll = () => {
     if (window.scrollY > 50) {
@@ -23,39 +23,13 @@ export default function Navbar() {
     }
     setLastScrollY(window.scrollY);
   };
+
   useEffect(() => {
     window.addEventListener("scroll", handleScroll);
     return () => {
       window.removeEventListener("scroll", handleScroll);
     };
   }, [lastScrollY]);
-
-  useEffect(() => {
-    const checkUser = async () => {
-      const {
-        data: { session },
-      } = await supabase.auth.getSession();
-      setUser(session?.user || null);
-    };
-    checkUser();
-
-    const { data: listener } = supabase.auth.onAuthStateChange(
-      async (event, session) => {
-        setUser(session?.user || null);
-      }
-    );
-
-    window.addEventListener("scroll", handleScroll);
-    return () => {
-      window.removeEventListener("scroll", handleScroll);
-      listener.subscription.unsubscribe();
-    };
-  }, [lastScrollY]);
-
-  const handleLogout = async () => {
-    await supabase.auth.signOut();
-    setUser(null);
-  };
 
   return (
     <nav className={`navbar ${isVisible ? "show" : "hide"}`}>
@@ -82,14 +56,16 @@ export default function Navbar() {
       <div className="nav-links">
         {user ? (
           <>
-            <Link
-              href="#"
-              className="login-link" // TODO: BYT KLASSNAMN KANSKE
-              onClick={(e) => {
-                e.preventDefault();
-                handleLogout();
-              }}
-            >
+            <Link href="/dashboard" className="nav-link">
+              Deltagare
+            </Link>
+            <Link href="/event" className="nav-link">
+              Eventet
+            </Link>
+            <Link href="/settings" className="nav-link">
+              Inställningar
+            </Link>
+            <Link href="#" className="nav-link" onClick={logout}>
               Logga ut
             </Link>
           </>
